Function ic_crypto_internal_bls12381_common::scalar_multiply

pub fn scalar_multiply<G: CurveProjective<Scalar = Fr>, F: Into<FrRepr>>(
    base: G, 
    factor: F
) -> G

Multiply an element of a group by a scalar. This can be applied to any group over Fr; in particular G1 and G2. The factor can be &FrRepr or &Fr. The difference is that FrRepr may be larger than the modulus. This is a pure wrapper around the impure function provided by the pairing library.